pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-commit"
version = "0.1.0"
description = "A CLI tool to generate commit messages using LLMs"
authors = [
    {name = "GarupanOjisan"}
]
dependencies = [
    "PyYAML>=6.0",
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "google-generativeai>=0.3.0",
    "requests>=2.0.0",
]
requires-python = ">=3.14"

[project.scripts]
llm-commit = "llm_commit.cli:main"
